+++ sql-plugin/src/main/scala/com/nvidia/spark/rapids/conditionalExpressions.scala
@@ -16,7 +16,9 @@
 
 package com.nvidia.spark.rapids
 
-import ai.rapids.cudf.{BinaryOp, ColumnVector, DType, NullPolicy, Scalar, ScanAggregation, ScanType, Table, UnaryOp}
+import java.util.function.Consumer
+
+import ai.rapids.cudf._
 import com.nvidia.spark.rapids.Arm._
 import com.nvidia.spark.rapids.RapidsPluginImplicits._
 import com.nvidia.spark.rapids.jni.CaseWhen
@@ -24,8 +26,9 @@ import com.nvidia.spark.rapids.shims.ShimExpression
 
 import org.apache.spark.sql.catalyst.analysis.{TypeCheckResult, TypeCoercion}
 import org.apache.spark.sql.catalyst.expressions.{ComplexTypeMergingExpression, Expression}
-import org.apache.spark.sql.types.{BooleanType, DataType, DataTypes}
+import org.apache.spark.sql.types._
 import org.apache.spark.sql.vectorized.ColumnarBatch
+import org.apache.spark.unsafe.types.UTF8String
 
 object GpuExpressionWithSideEffectUtils {
 
@@ -338,7 +341,7 @@ case class GpuCaseWhen(
   }
 
   private lazy val useFusion = caseWhenFuseEnabled && branches.size > 2 &&
-    GpuColumnVectorUtils.isCaseWhenFusionSupportedType(inputTypesForMerging.head) &&
+    isCaseWhenFusionSupportedType(inputTypesForMerging.head) &&
     (branches.map(_._2) ++ elseValue).forall(_.isInstanceOf[GpuLiteral])
 
   override def checkInputDataTypes(): TypeCheckResult = {
@@ -387,7 +390,7 @@ case class GpuCaseWhen(
               .asInstanceOf[GpuScalar])
           withResource(thenElseScalars) { _ =>
             // 2. generate a column to store all scalars
-            withResource(GpuColumnVectorUtils.createFromScalarList(thenElseScalars)) {
+            withResource(createFromScalarList(thenElseScalars)) {
               scalarCol =>
                 val finalRet = withResource(new Table(scalarCol)) { oneColumnTable =>
                   // 3. execute final select
@@ -584,5 +587,99 @@ case class GpuCaseWhen(
     val cases = branches.map { case (c, v) => s" WHEN ${c.sql} THEN ${v.sql}" }.mkString
     val elseCase = elseValue.map(" ELSE " + _.sql).getOrElse("")
     "CASE" + cases + elseCase + " END"
+  }
+
+  private def isCaseWhenFusionSupportedType(dataType: DataType): Boolean = {
+    dataType match {
+      case BooleanType => true
+      case ByteType => true
+      case ShortType => true
+      case IntegerType => true
+      case LongType => true
+      case FloatType => true
+      case DoubleType => true
+      case StringType => true
+      case _: DecimalType => true
+      case _ => false
+    }
+  }
+
+  /**
+   * Create column vector from scalars
+   *
+   * @param scalars literals
+   * @return column vector for the specified scalars
+   */
+  private def createFromScalarList(scalars: Seq[GpuScalar]): ColumnVector = {
+    scalars.head.dataType match {
+      case BooleanType =>
+        val booleans = scalars.map(s => s.getValue.asInstanceOf[java.lang.Boolean])
+        ColumnVector.fromBoxedBooleans(booleans: _*)
+      case ByteType =>
+        val bytes = scalars.map(s => s.getValue.asInstanceOf[java.lang.Byte])
+        ColumnVector.fromBoxedBytes(bytes: _*)
+      case ShortType =>
+        val shorts = scalars.map(s => s.getValue.asInstanceOf[java.lang.Short])
+        ColumnVector.fromBoxedShorts(shorts: _*)
+      case IntegerType =>
+        val ints = scalars.map(s => s.getValue.asInstanceOf[java.lang.Integer])
+        ColumnVector.fromBoxedInts(ints: _*)
+      case LongType =>
+        val longs = scalars.map(s => s.getValue.asInstanceOf[java.lang.Long])
+        ColumnVector.fromBoxedLongs(longs: _*)
+      case FloatType =>
+        val floats = scalars.map(s => s.getValue.asInstanceOf[java.lang.Float])
+        ColumnVector.fromBoxedFloats(floats: _*)
+      case DoubleType =>
+        val doubles = scalars.map(s => s.getValue.asInstanceOf[java.lang.Double])
+        ColumnVector.fromBoxedDoubles(doubles: _*)
+      case StringType =>
+        val utf8Bytes = scalars.map(s => {
+          val v = s.getValue
+          if (v == null) {
+            null
+          } else {
+            v.asInstanceOf[UTF8String].getBytes
+          }
+        })
+        ColumnVector.fromUTF8Strings(utf8Bytes: _*)
+      case dt: DecimalType =>
+        val decimals = scalars.map(s => {
+          val v = s.getValue
+          if (v == null) {
+            null
+          } else {
+            v.asInstanceOf[Decimal].toJavaBigDecimal
+          }
+        })
+        fromDecimals(dt, decimals: _*)
+      case _ =>
+        throw new UnsupportedOperationException(s"Creating column vector from a GpuScalar list" +
+            s" is not supported for type ${scalars.head.dataType}.")
+    }
+  }
+
+  /**
+   * Create decimal column vector according to DecimalType.
+   * Note: it will create 3 types of column vector according to DecimalType precision
+   *  - Decimal 32 bits
+   *  - Decimal 64 bits
+   *  - Decimal 128 bits
+   *    E.g.: If the max of values are decimal 32 bits, but DecimalType is 128 bits,
+   *    then return a Decimal 128 bits column vector
+   */
+  private def fromDecimals(dt: DecimalType, values: java.math.BigDecimal*): ColumnVector = {
+    val hcv = HostColumnVector.build(
+      DecimalUtil.createCudfDecimal(dt),
+      values.length,
+      new Consumer[HostColumnVector.Builder]() {
+        override def accept(b: HostColumnVector.Builder): Unit = {
+          b.appendBoxed(values: _*)
+        }
+      }
+    )
+    withResource(hcv) { _ =>
+      hcv.copyToDevice()
+    }
   }
 }
